# docs/api-swagger.yaml
openapi: 3.0.3
info:
  title: Dzinza Genealogy Platform API
  version: 1.0.0
  description: |
    This is the OpenAPI (Swagger) specification for the Dzinza Genealogy Platform API.
    It covers authentication, user management, and core genealogy features.
servers:
  - url: http://localhost:3001/api
    description: Local development server
paths:
...
  /auth/register:
    post:
      summary: Register a new user
      tags: [Authentication, Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - firstName
                - lastName
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
                firstName:
                  type: string
                lastName:
                  type: string
                preferredLanguage:
                  type: string
                  enum: [en, sn, nd]
                  default: en
      responses:
        "201":
          description: User registered successfully
        "400":
          description: Validation error
        "409":
          description: Email already exists
  /register:
    post:
      summary: User registration
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                full_name:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: Registration successful
  /auth/login:
    post:
      summary: Login a user
      tags: [Authentication, Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials
  /login:
    post:
      summary: User login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Token refreshed
        "401":
          description: Invalid or expired refresh token
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []